+++ CHANGELOG.md
@@ -43,11 +43,13 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 
 #### :bug: Bugfixes
 * Fix rendering of HTML content in KeepRight and ImproveOSM issues ([#8928])
+* Fix "local storage is exceeded" error message showing up multiple times ([#8930])
 #### :hammer: Development
 * Switch build system to [esbuild](https://esbuild.github.io/) for much faster builds ([#8774], thanks [@ mbrzakovic] and [@bhousel])
 
 [#8774]: https://github.com/openstreetmap/iD/pull/8774
 [#8928]: https://github.com/openstreetmap/iD/pull/8928
+[#8930]: https://github.com/openstreetmap/iD/pull/8930
 
 
 # 2.20.3